+++ src/components/Logo.jsx
@@ -1,9 +1,10 @@
 import { Box } from "@mui/material";
 import Image from "next/image";
 import Link from "next/link";
+import { memo } from "react";
 
 // ゴミ処理　面倒でLogo以外の要素も持たせてある
-export const Logo = () => {
+export const Logo = memo(function Logo() {
   return (
     <Box sx={{ px: 3 }}>
       <Link href="/" passHref>
@@ -13,4 +14,4 @@ export const Logo = () => {
       </Link>
     </Box>
   );
-};
+});
