+++ src/components/Navbar.jsx
@@ -2,20 +2,44 @@ import { FilterAltOutlined, SearchOutlined } from "@mui/icons-material";
 import { AppBar, Box, IconButton, Toolbar, Tooltip } from "@mui/material";
 import { styled } from "@mui/material/styles";
 import { useRouter } from "next/router";
-import { memo } from "react";
+import { memo, useCallback } from "react";
+import { useDispatch } from "react-redux";
 
 import { Logo } from "@/components/Logo";
+import {
+  toggleSearchOpen,
+  toggleSelectionOpen,
+} from "@/modules/features/app/slice";
 
 const NavbarRoot = styled(AppBar)(({ theme }) => ({
   backgroundColor: theme.palette.background.paper,
   boxShadow: theme.shadows[3],
 }));
 
-export const Navbar = memo(function Navbar({
-  handleToggleSearchOpen,
+const SelectionTooltip = memo(function SelectionTooltip({
   handleToggleSelectionOpen,
 }) {
   const router = useRouter();
+
+  return router.pathname === "/people/[personId]" ? (
+    <Tooltip title="フィルター" sx={{ display: { lg: "none" } }}>
+      <IconButton onClick={handleToggleSelectionOpen}>
+        <FilterAltOutlined />
+      </IconButton>
+    </Tooltip>
+  ) : null;
+});
+
+export const Navbar = memo(function Navbar({}) {
+  const dispatch = useDispatch();
+
+  const handleToggleSearchOpen = useCallback(() => {
+    dispatch(toggleSearchOpen());
+  }, [dispatch]);
+
+  const handleToggleSelectionOpen = useCallback(() => {
+    dispatch(toggleSelectionOpen());
+  }, [dispatch]);
 
   return (
     <NavbarRoot>
@@ -31,13 +55,9 @@ export const Navbar = memo(function Navbar({
       >
         <Logo />
         <Box>
-          {router.pathname === "/people/[personId]" && (
-            <Tooltip title="フィルター" sx={{ display: { lg: "none" } }}>
-              <IconButton onClick={handleToggleSelectionOpen}>
-                <FilterAltOutlined />
-              </IconButton>
-            </Tooltip>
-          )}
+          <SelectionTooltip
+            handleToggleSelectionOpen={handleToggleSelectionOpen}
+          />
           <Tooltip title="検索" sx={{ display: { lg: "none" } }}>
             <IconButton onClick={handleToggleSearchOpen}>
               <SearchOutlined />
